@@ -13,22 +13,35 @@
     "@xstate/vue": "^2.0.0",
     "awesome-snackbar": "^2.0.7",
     "axios": "^1.5.1",
-    "vue": "^3.2.45",
+    "path": "^0.12.7",
+    "vue": "^3.3.9",
     "vue-connect-wallet": "^0.1.7",
     "vue-google-oauth2": "^1.5.10",
+    "vue-i18n": "^10.0.4",
     "vue-router": "^4.1.6",
     "vue3-google-login": "^2.0.23",
     "xstate": "^4.35.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.12.0",
+    "@types/node": "^22.7.6",
+    "@typescript-eslint/eslint-plugin": "^6.13.0",
     "@vitejs/plugin-vue": "^4.0.0",
     "@vueuse/core": "^9.10.0",
     "@vueuse/motion": "2.0.0-beta.12",
     "@xstate/cli": "^0.4.1",
     "autoprefixer": "^10.4.13",
+    "eslint": "^8.57.1",
+    "eslint-config-standard-with-typescript": "^40.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-n": "^16.3.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.29.0",
+    "globals": "^15.11.0",
     "postcss": "^8.4.20",
     "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.3",
+    "typescript": "^4.9.5",
+    "typescript-eslint": "^8.9.0",
     "vite": "^4.0.0",
     "vite-plugin-pages": "^0.28.0",
     "vite-svg-loader": "^4.0.0",
