@@ -1,5 +1,6 @@
-import { createRouter, createWebHistory } from "vue-router";
+// @ts-expect-error: Vue router routes import from directory structure
 import routes from "~pages";
+import { createRouter, createWebHistory } from "vue-router";
 
 const router = createRouter({
   history: createWebHistory(),
