@@ -1,11 +1,7 @@
-/// <reference types="vite-svg-loader" />
-/// <reference types="vite-plugin-pages/client" />
-
-declare module '*.vue';
+declare module "*.vue";
 
 declare module "*.svg" {
   import { DefineComponent } from "vue";
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
-  const component: DefineComponent<{}, {}, any>;
+  const component: DefineComponent<object, object, object>;
   export default component;
 }
